Bug#28836474 EVENT TE_ALTER DROPPED IN NDBAPI DUE TO EARLY SUB_GCP_COMPLETE_REP FROM SUMA
Problem
=======

The SUB_GCP_COMPLETE_REP signal is used to signal to receiver that an epoch
is completed and that receiver have received all data for the epoch.

For alter table a fragmented signal (SUB_TABLE_DATA event TE_ALTER) may be
sent from SUMA to NdbAPI.

Since it is the arrival of the last fragment in the signal that determines
when the signal as a whole is delivered, fragmented signals typically are
delivered out of order with respect to when first fragment is sent.

SUMA do not keep track on fragmented signals in flight, and when SUMA have
received all SUB_GCP_COMPLETE_REP signals for an epoch it assumes all data
have been sent to all NdbAPI nodes and directly sends SUB_GCP_COMPLETE_REP
to all NdbAPI nodes.

Now for a NdbAPI node all the SUB_GCP_COMPLETE_REP signals for an epoch may
arrive before the last fragment of the fragmented signal (SUB_TABLE_DATA).

When the fragmented signal eventually is delivered the epoch is already
marked closed and the SUB_TABLE_DATA signal is ignored.

Solution
========

Delay send of SUB_GCP_COMPLETE_REP until all fragments in fragmented
signals for an epoch are sent to all API nodes.