Bug#28888936 CREATING HISTOGRAM STATISTICS FOR LARGE DATA SETS IS SUSPICIOUSLY SLOW
The class Value_map acts as a container which counts the number
of occurrences for each distinct value. The container is also
sorted by the value, and the use case is for creating histogram
statistics. The Value_map originally used
std::vector<std::pair<T, ha_rows>> as the internal container, and
this was kept sorted by inserting values to the correct position.
This worked well for data sets that had few distinct values, but it
became very slow for data sets with many distinct values.

This patch switches the internal container from
std::vector<std::pair<T, ha_rows>> to std::map<T, ha_rows>. The
performance for data sets with few distinct values is marginally
slower, but for larger data sets with many distinct values it is
significantly faster. The downside is that the memory overhead for
std::map is bigger, but that is neglible compared to the performance
gains.

The performance gain is also seen when generating a histogram
for a BIGINT column with approx. 67 million rows. Before this patch,
it took over seven minutes to complete on my local machine. After
this patch, it completes in around 36 seconds (though with a lower
sampling percentage due to higher memory overhead).

Microbenchmarks (Haswell 3.4 GHz, optimized mode, GCC 8.2):

  BM_ValueMapInsertionIntSmallRange     93000 ->  100020 ns/iter [  -7.0%]
  BM_ValueMapInsertionIntMediumRange   247635 ->  178815 ns/iter [ +38.5%]
  BM_ValueMapInsertionIntLargeRange    676973 ->  207255 ns/iter [+226.6%]
  BM_ValueMapInsertionString          1429669 -> 1164892 ns/iter [ +22.7%]

Change-Id: I22b9bf707c6ef8a9e6981aa9b32ba3e7ad0a712c