Bug#27429615 CLUSTER: QUERY EXPLAIN SHOWS PUSHED CONDITIONS WHICH ARE NOT PUSHED
Preparation patch:

Move the generate_scan_filter() call for pushed joins from where we build
a NdbQuery definition, to where the NdbQuery operations are instantiated
from the definition.

The effect is that generate_scan_filter() for pushed joins is now performed
at the same logical place in the code as for non pushed NdbOperation.
(Now done together with instantiating the NdbOperations to be executed)

This is needed due to later patches moving the call to handler::cond_push()
to an execution point to after any pushed join has been created.

Changes in several ndb_join_pushdown_*.result files, where the
'scan filter too large' warning is not emitted as part of
*explain* of queries. This is due to generate_scan_filter not
being executed as part of optimize phase of the query. Note that
there are still such warnings when the same query is executed.