Bug#28877215 CANNOT CREATE A JSON VALUE FROM A STRING WITH CHARACTER SET 'BINARY'
When trying to select data from a JSON column in the FEDERATED
engine, the server would return the error "Cannot create a JSON value
from a string with CHARACTER SET 'binary'.". The reason is that when the
server is reading data from the external table, it is storing the data
into the local fields using the character set my_charset_bin regardless of
the character set of the external table/fields. Furthermore, Field_json::store
expects all incoming data to be in the character set utf8mb4_bin. So
when calling Field_json::store, the server returned an error. This patch
adds a special case for JSON fields where the character set is overridden
to be utf8mb4_bin when storing data from external fields.

This patch also fixes a problem where neither DELETE or UPDATE had any
effect if the table contained a JSON column. As an example: When doing
"UPDATE tbl SET col1 = 1;" on a federated table, the federated storage
engine builds a SQL query on the form "UPDATE tbl SET col1 = 1 WHERE
col1 = value AND ... AND coln = value" where all columns are included
in the WHERE clause. This query is sent back to the original source
table. For JSON columns, the storage engine did not wrap the value in
a CAST(... AS JSON), so a match would never be found in the original
source table (comparison between a string value and a JSON value will
never return true) This patch adds an explicit CAST(... AS JSON) around
all JSON columns for both UPDATE and DELETE statements done in the
federated engine.

Change-Id: I9ea63a873af2f336ffb7a72b9e440d1dc958965c