Bug#28897992 NDB FK DATATYPE AND VAR NAME FIXUP
The ha_ndb_ddl_fk.cc file uses variables named "no" for describing an
index value, this can easily be interpreted as "not yes".
Also the variable is not marked as const even though it's only assigned
once.
Furthermore the variable is using wrong data type, both the function
which assigns the variable and the functions called uses "int".

Fix by not using variable name "no",  change to datatype "int" and mark
the variable as "const".

Change-Id: I95ed0c9f51c7d86314d3fea7085f80062af418cc