Bug #28777704 SHOW FUNCTION STATUS" ASSERTION FAILED: !TABLE || (!TABLE->READ_SET..... Bug #28769996 ASSERT !TABLE || (!TABLE->READ_SET || BITMAP_IS_SET(TABLE->READ_SET, FIELD_INDEX
The assert ASSERT_COLUMN_MARKED_FOR_READ was hit in Field_enum::val_int()
when comparing two enum fields using the method Field_enum::cmp().

This assert is intended to check that mysqld does not read back
any column values it has not explicietely set in the table::read_set
bitmap. These values are then read with the Field::val_*() methods.

However, in the cases reported in the bugs referred above, the
temptable handler is retrieving the field values as part of
a handler internal operation where two keys are being compared.
Thus, the requested field values are not to be returned to the
server as part of this operation.

This situation could typically be encountered in the generic
query:

  Select a from <table> where b = 'x'

If we have an index on column b, the condition 'b = 'x'' will
cause a REF access to be selected, and the entire condition will
be eliminated as it is covered by the REF('x')  access. As there is
no condition left to be evaluated by the server, only column a will
be present in the read_set.

However, internally in the handler we will have to do compare operations
against column 'b'. Both in order to navigate to the first row containing
b = 'x', and checking when 'read_next' has reached the end of range where
b = 'x'.

It will vary between handler implementations how this is handled; Some may
use memcmp() the binary data, other push the entire range check down to the
storage engine (NDB), and some use the Field::cmp*() methods for this
(temptable)

Looking at the different implementations of the Field::cmp() methods,
we find that none of these enforce the ASSERT_COLUMN_MARKED_FOR_READ
as these are intended/allowed for accesing field values without
being part of the read_set.

However, for Field_enum datatypes, the method ::int_val() is used as
a convenienet way to retrieve the values to be compared. Here we
may hit the ASSERT_COLUMN_MARKED_FOR_READ!

This patch introduce the static enum_int_val() utility function
with logic similar to Field_enum::int_val(), *without* the
ASSERT_COLUMN_MARKED_FOR_READ.

Field_enum::cmp() and Field_enum::int_val() is changed to use
this function to return the enum int value.