Bug#27556117: SIG11 IN HANDLER::HA_STATISTIC_INCREMENT WITH CLONED DATADIR+LOAD
Analysis:
========
The reported behavior is observed due to race condition on
THD::m_transaction in binlog group commit.  The group commit
leader thread will inadvertently start an attachable
transaction as part of executing ha_commit() for a follower
thread. The temporary installation of the Attachable_trx in
THD::m_transaction then ends up terminating the condition
variable wait loop (in bool Stage_manager::enroll_for()),
which keeps the follower threads blocked while the leader is
committing on their behalf. As a result both the leader and
follower threads executes in the same THD. In debug mode
this often results in the DBUG_ASSERT() immediately
following the call to Stage_manager::enroll_for() crashing
because the cache_manager pointer in the THD is temporary
cleared by the attachable transaction.

Fix:
====
Instead of THD::m_transaction a new thread specific bool
variable named 'tx_commit_pending' is introduced to track if
transaction commit is pending or complete.