Bug#28550736 RACE CONDITION ON VIEW_CHANGE_NOTIFIER
Since multiple pieces of code now call and use the current view
changing notification service, its unique and centralized nature was
causing minor issues.
Such issues included waiting and timing out for a view that was
already delivered on error branches.

This code was now refactored, and when waiting on a view you use a
local object registered in a central list.

Other minor improvements in terms of logging were added.