WL#12668: Pre-parse plugin for DDL statement clause filtering.
This patch implements a plugin to do pre-parse regex based
replacement of CREATE TABLE clauses. The following replacements
are implemented:

"DATA|INDEX DIRECTORY [=] '<path>'" ==> " "
"ENCRYPTION           [=] 'N|Y'"    ==> " "

The plugin is intended used when restoring dumps into environments
where the clauses listed above are not desirable.

It should be noted that due to the nature of the plugin, being
based on regex replacement, there may be false positives.

Change-Id: I278259250a0c816872bc21ec6817f3fa3e97d9dd