Bug#28547906 ENUM TYPE CASTING WORKS WRONG WITH SUBQUERIES
Problem:

The method `Item::cast_to_int_type' is used in the non-nested query case
by `Item_func_(un)signed::val_int' to determine whether to call `val_int'
or not.  For the case of an enum, in the non-nested case, we called
the overloaded `Item_field::cast_to_int_type', which returns INT for an
enum.  For the subquery case, we have an interventing `Item_view_ref',
which does not have an overload of `cast_to_int_type', so MySQL resorts to
the default `Item::Item_field::cast_to_int_type', which returns an
answer base on the item's result type, which in this case is
STRING_RESULT, so we end up calling `val_str' instead of `val_int', hence
the warnings.

Solution:

Define an overload Item_ref::cast_to_int_type, which forwards the call
the referenced item.

Change-Id: I0d76d580a7da34b326e2aaba9d7dc7d1b33fd464