Bug#11748067: FEDERATED TABLES BLOB FIELD IS NOT CORRECTLY UPDATED
When a BLOB column in the Federated storage engine is updated, the
updated value overwrites the copy of the old column value. The
executor therefore thinks that the update is a no-op, and the updated
value is not stored in the table.

The fix is to make the Federated storage engine copy the BLOB data to
a separate memory area, outside of the Field object. This way, when
the Field object is updated with the new BLOB value, it does not
overwrite the copy of the old data.

(The fix is copied from ha_tina, which already did this correctly,
with some adjustments such as using helper functions for manipulating
the BLOB pointers in the record buffer. The adjustments are also
applied to ha_tina to keep the code in the two engines as similar as
possible.)

Change-Id: I7e877759bbd8404be98829d8347237f60253ff91