Bug#28536312 - MYSQL SERVER CRASH WHEN ROLLBACK A                'CREATE USER' STATEMENT
DESCRIPTION:
============
When a 'create user' statement is rolled back, there is
a server crash because of an ASSERTION failure.

ANALYSIS:
========
When a rollback happens, the plugin decides whether to abort
the transaction and then set an error on the thread.
However, this error is ignored and not correctly propagated
to the calling functions because of which mysql_create_user
finally returns success. Later on, the server crashes
in my_ok where it expects that there should not have been
any error.

FIX:
====
The fix makes sure that the error is correctly captured.