Bug#28538971 - ASAN: DIRECT LEAK OF 272 BYTE(S) IN                MAIN.MYSQLPUMP_PARTIAL_BKP MTR TEST CASE
DESCRIPTION:
===========
When mysqlpump encounters an error, it does not free all
allocated resources resulting in memory leaks.

ANALYSIS:
========
In normal scenario, when a dump task is being processed,
it is added to the vector which gets cleaned up at the end
of the execution. However, when an error is encountered,
we simply return from the processing function without
keeping a track of those tasks. This doesn't let
cleanup to happen properly and results in memory leak.

FIX:
====
Added code to clean up the resources in case of errors.