Bug #28862343 ATRT SUCCEED DESPITE CRITICAL ERRORS
Fixes the error reporting of ATRT in two cases.

ATRT was reporting success when wrong parameters are passed and when any of the
processes weren't stopped after all the tests ran. Now, return_code is set to
ATRT_FAILURE, in both the above cases making it a critical error and failing
the ATRT.