Bug #28860795	ERRMSG-UTF8.TXT USES WRONG FORMAT SPECIFIERS
The format specifiers used in errmsg-utf8.txt did not correspond correctly to actual types passed to ib::info, ib::warn, ib:error and ib::fatal.
This issue was not very visible because the format strings are not available to compiler at compile time, types of integers dependent on many ifdef/typedefs which were platform specific, and the issue manifested only for really large numbers on certain platforms.
This patch:
- adds a run-time verification of datatypes in debug build
- fixes places where format specifier were wrong
- fixes some obvious typeos in some error messages
- fix for one place where string was not null terminated
- an MTR for the above fix

This patch also observes quite simple rules, which prevent typing problems:

1. When calling ib::info/warn/error/fatal pass arguments which have types which directly correspond to format specifiers:

%lld  - long long int (a.k.a. longlong)
%llu, %llx - unsigned long long int (a.k.a. ulonglong)
%ld - long int (a.k.a. long)
%lu, %lx, %lX - unsigned long int (a.k.a. ulong)
%lf - double
%d, %i - int
%u, %x, %X - unsigned int (a.k.a. uint)
%f - float
%c - char
%p - a pointer
%s - char *, char const *, char[N], or any other thing which points to array of chars
%zu, %zx - size_t
%zd - ssize_t

2. When the data you want to print is not guaranteed to have one of the types listed above, convert it using an initializer list:

	ulonglong{start_lsn}i

as oposed to using static_cast<ulonglong>, this has an advantage that compiler will warn you if on your platform the destination type can not hold the source type (a.k.a. narrowing).

Reviewed-by: Sunny Bains <Sunny.Bains@oracle.com>