Bug#28947107: JSON_OBJECT() CASTS YEAR VALUES TO BASE64
YEAR values were stored as opaque data in JSON, and when JSON
documents with YEAR values were converted to text, the YEAR values
were shown as base64-encoded strings.

Now YEAR values are stored as unsigned integers instead. This is more
compact in terms of storage space, and it makes the years come out as
plain numbers instead of base64-encoded strings when the JSON document
is converted to text.

Change-Id: I7ea4ca1a87b5b4d52fdcc6de78de782d48da9a78