Bug#27309336 Backport to 5.7
Problem:
Hang observed while attempting to create large number of tables as fast
as possible if innodb_flush_method is set to O_DIRECT_NO_FSYNC. No hang
or crash is observed when the innodb_flush_method is set to O_DIRECT.
The problem seems to be because of out of sync FS meta data as fsync()
is never called if the flush method is set to O_DIRECT_NO_FSYNC.

Fix:
Call fsync in O_DIRECT_NO_FSYNC mode whenever the size of the file changes.
This will keep the FS metadata in sync.

RB: 20994
Reviewed by: Debarun Banerjee (debarun.banerjee@oracle.com)