Bug #25960230 899 rowID ALREADY ALLOCATED ERROR UNDER HEAVY LOAD
Error 899 "rowID already allocated" occurs when a rowID is
available at the primary fragment replica and unavailable at a
backup fragment replica. An insert operation seizes the rowID
at the primary and fails to seize it at the backup since it is
unavailable. This causes transaction failure, forcing
applications to perform retries which are inconvenient and
reduce performance.

One common scenario is when there are two transactions operating
on rows in the same fragment - one transaction which deletes a
row,and a second transaction which is inserting rows into the same
fragment. The first transaction is releasing a rowID into the
'freerowIDs pool' of the fragment, and the second is seizing a
rowID from that pool. It is important that any rowID seized from
the 'free rowIDs pool' on the primary fragment replica is available
on all other fragment replicas, to avoid hitting error 899.

The design to ensure this involves releasing the rowID back to the
pool when both:

-ACC no longer has any prepared operations referencing the tuple
-The last operation in the deleting transaction completes on each
replica.

As complete processing occurs on backup replica(s) first, it
should be guaranteed that any rowID released on the primary replica
has previously been made available on the backup replica(s).

However the current implementation uses the commit order of
operations within a transaction, rather than the complete order, and
in some cases these are not the same, resulting in rowIDs being
released to the pool on the Primary fragment replica too early,
resulting in transient error 899 conditions for other transactions.

DESIGN DETAILS

For multiple operations on the same tuple where the end result of
the transaction is that the tuple is deleted, the earlier implementation
was that the last operation to commit on that tuple was marked as the
deallocator of the rowID. The operation was marked by setting the
m_dealloc flag during COMMIT. When any op COMPLETEd, its m_dealloc
flag was checked, and if set, its rowID was deallocated.

A problematic case was when a transaction T1 commited/completed in the
unexpected (not incorrect) order:

COMMIT(DELETE)->COMMIT(READ)->COMPLETE(READ)->COMPLETE(DELETE)

(Note that the COMMIT and COMPLETE orders are different
COMMIT : DELETE, READ
COMPLETE : READ, DELETE)

-T1: DELETE commits on primary
-T1: READ commits on primary, m_dealloc=1 for READ on primary
-T1: READ completes on primary, deallocates rowID X:Y
-T2: INSERT seizes freed rowID X:Y on primary
-T2: INSERT tries to seize rowID X:Y on backup, gets error 899 since
rowID is not freed yet on backup
-T1: DELETE completes on backup, deallocates rowID X:Y

The different COMMIT and COMPLETE orders were because the signals for
committing/completing the single-replica READ and the all-replicas
DELETE took different paths, and had non-deterministic delivery orders.
Mixing locking reads and DELETEs is a very common access pattern.

Solving this problem is made more complex by the spread of operation
state across ACC, LQH and TUP.
-ACC contains a good overview of the cumulative effect of multiple
operations on a key, but ACC also considers further use of the key in
subsequent transactions using a different rowID, and does not retain any
state after COMMIT.
-TUP owns the rowID, but has fairly optimised and change-sensitive
(tuple) data structures which are tricky to modify to store more meta
state. Also TUP is not currently involved in COMPLETE processing other
than for rowID deallocation.
-LQH connects ACC and TUP, is involved in COMPLETE processing, and has
existing per operation state which is retained until COMPLETE time.

At COMMIT time, a reference count is stored by LQH in a designated 'counting
op' in the ACC op queue. All other ops in the queue store the ID of the
'counting' op in LQH, so the reference count is accessible for each
COMMIT and COMPLETE. The reference count is incremented for every
operation involved in the transaction commit when the first COMMIT
arrives, and decremented as each COMPLETE is processed. The rowID is
deallocated when the reference count reaches 0. This guarantees that the
dealloc will not happen until the last op COMPLETEs, on each replica so
early dealloc due to differing replica COMMIT and COMPLETE orders cannot
occur.

IMPLEMENTATION DETAILS

The 'counting' op must be accessible until the last op COMPLETEs because
it holds the reference count. If the 'counting' op is not the last op to
COMPLETE, it is marked as a 'zombie' op, and released together with the
rowID when the reference count reaches zero. The existing ACC code does
not trigger rowID release while any scanning operation has a lock. This
can occur when

e.g. :
-A locking scan finds a batch of rows which are candidates for deletion
-PK (takeover) ops are issued to delete them in a separate transaction
-Separate transaction is committed
-Scan batch is unlocked and moves forward

For this case, the existing code will retain the rowID until the scan
op(s) commit, then perform an immediate dealloc, which may be too soon
if the complete of the deleting ops is still occurring.

To handle this with reference counting, an additional logical increment
is always performed on the reference count. ACC provides the extra
decrement when the last op commits, passing to LQH a reference to the
'counting' op which was stored in the ACC scan operation record. This
allows ACC to 'hold up' deallocation while waiting for (LQH-stateless)
scan operations to COMMIT, and also solves any problems with the state
of still-completing DELETE ops.

Allowing ACC to hold up deallocation requires that ACC behaves well wrt
follow-on transactions on the same keys

E.g. that :
-They do not use the same rowID, but seize a fresh one
-They do not get involved in the row deletion state and reference counting

One problem discovered here was the case mentioned in bug#19031389 where
a transaction containing a mix of read-sh and delete commits the delete
in ACC before the read-sh - ACC then treats the overall key state as
read-sh and allows following transactions' ops to read the key (and
immediately fail as it is deleted). This was problematic as the
following transaction's operations were not involved in reference
counting, but stopped ACC from indicating that all operations on the
rowID were complete.

To avoid this, the fix for bug#19031389 was changed to not allow a
following transaction's operations to execute until all operations
involved in a deleting transaction have committed at ACC (and the ACC
side of reference counting is finished).

There are no modifications to the ABORT of INSERT behaviour. As normal,
a single op (the INSERTing op) is designated as the 'counting' op. The
reference count is incremented to one for the 'counting' (INSERT) op
abort processing at ACC, and then immediately decremented again, so that
LQH then instructs TUP to deallocate the op immediately when releasing
the LQH operation. ABORT processing may have its own 899 issues, but
those are out of scope of this fix.

TESTCASES ADDED/USED IN VERIFYING THIS FIX

1) Added ndb_err899.test: transactions with deletes run concurrently
with an insert transaction where COMMIT/COMPLETE orders are scrambled.

-single delete in delete transaction
-multiple deletes in delete transaction
-multiple deletes + concurrent reads
-multiple deletes + abort

2) Used autotest 'testIndex -r 10 -n NF_Mixed T1 T6 T13' for node-
failure scenarios

3) Used autotest 'testBasic -n AccCommitOrder -v T1' for commits+aborts