Bug#28864244 - DATA DICTIONARY ASSERT IN MEB.UNICODE
While creating MDL key for routines, events and triggers, debug
assert check is added to verify parameter schema name is in
lowercase when lower_case_table_names == 2. This assert condition
fails for the schema name with a multi-byte character (for example:
tèst). On non-debug execution continues without any issues.

To check schema name is in lowercase, now we check if any of its
character in uppercase by calling my_isupper(). Assert condition
fails if any character is in uppercase. my_isupper() and
other similar macros are more suitable with one-byte character
sets but with multi-byte character (as 'è') the method returned
incorrect result and assert condition failed.

To fix this issue, now changed code to prepare schema name with
lower case and memcmp it with the actual schema name in the
assert condition.