Bug #28956360: CRASH WHEN CALLING MYSQL_REAL_CONNECT() IN LOOP
When uninitializing and reinitialized the client library, character sets
(which would only be initialized if we actually had done a connect) would not
be properly reinitialized on the second round, as std::call_once really only
runs once.

Resetting std::once_flag is not strictly supported by the standard, but there
seems to be consensus that doing it with placement new is fine, as long as no
other threads try to access it at the same time (which should definitely be
the case when doing shutdown of the library).

Change-Id: I32d4e143474160e6ad92fbb9f82638bc7d3adcf5