Bug#28961997: REMOVE PARENTHESES INFORMATION FROM AST
This patch removes some hacks that were needed before we had
a single correctly written production rule for query
expressions. Back when the AST was created directly by the
parser, it needed to remember some nesting information that
it accidentally threw away. It did so by recording if it saw
parentheses, and then took various actions based on this.

This patch removes all notions of parentheses in the AST,
instead basing all decisions on the structure of the parse
tree, thereby separating the concerns: the parser takes care
of nesting, while the AST builder checks semantics.

The following areas are changed:

- The contextualization order by/limit clauses based its
  decisions partly on the already-build AST, partly on
  parenthesis information, and partly on the parse
  tree. This has now been changed to only pay attention to
  the parse tree, building the AST accordingly.

- Semantic checking for order by/limit is now done in the
  contextualization procedure, not by the AST.

- Printing of queries in EXPLAIN now always prints query
  expressions within unions with their own set of
  parentheses, irrespective of whether the original query
  used them or not. This accounts for all mtr test changes.

- In SELECT_LEX::setup_order_final(), the test for 'braces'
  is irrelevant, as queries like ((SELECT * FROM t1) ORDER
  BY t1.a DESC) LIMIT 2; fail anyway.

- In JOIN::optimize(), what the test really means is to rule
  out UNION. For e.g. (SELECT sql_calc_found_rows a FROM t1
  limit 1) UNION ALL SELECT a FROM t1, the rows in both
  query expression are counted.

Change-Id: I7fcabe2b61024e6882688bb8fb8602925921a59d