Bug#28705511: HEAP CURRUPTION WHEN STARTING MYSQLD WITH MANY TABLES
During start-up, we use a mechanism called "buffered logging"
which holds on to error log information until we know where to
log it to etc.  Due to an implementation detail of the component
framework, the "buffered logging" phase now ends after InnoDB
(which is multi-threaded) is initialized, where previously,
buffered logging ended before we started InnoDB initialization
(and thus, while we were still running single-threaded).
As the error-log backlog buffer -- in contrast to regular
log-events -- is globally shared, and InnoDB might try to log
from multiple threads at once, we must now take special care
to make it safe for it (and other multi-threaded code) to do so.