Bug#25115785 - CLUSTER DOESN'T START WITH MORE THAN 18 UNDO LOGS
Problem :
Summary: Incorrect count maintenance results in timeouts,
watchdogs taking nodes down due to lack of progress, or hang.

Details: When NDBFS receives a bound file operation request, it queues
the job to a shared bound queue. If there is no idle IO thread in the
bound pool to execute the request, it will create a new IO thread,
based on the count 'm_active_bound_threads_cnt'.

An IO thread from the pool dequeues a job from the queue and executes
it. An IO thread executing a bound openFileRequest removes itself from
the pool, attaches itself to the file being opened and increments the
count. It will then handle all the operations on the file. Only when
the file is closed, it will detach itself from the file, update the
count and return to the pool.

However, the count is not actively maintained for the other bound file
requests, such as, index build, allocate memory and remove file.

There is a lag between NDBFS producing a job to the queue and an IO
thread consuming it from the queue. Due to the lag, NDBFS may
overqueue jobs without creating new threads to handle them.

When most of the IO threads are attached to the files kept opened for
a long period, there may not be enough IO threads available to serve
the queued requests within an acceptable time bound.

This will result in timeouts, watchdogs taking nodes down due to lack
of progress, or hang (as observed in this bug report).

Fix:

Move the bookkeeping of the count from IO threads to NDBFS in order to
remove the lag and include all bound operations.

Note : Maintaining the counts in NdbFS from *before* a
request is queued until *after* it is completed means that idle
threads which have completed requests are not available for further
requests until NdbFS has received and processed their responses.

This lag should be fairly short as a wakeup socket is used to wake
NdbFs when a response is available.

Any existing code relying on threads more 'quickly' becoming available
for new requests is timing based and therefore likely to be
unreliable.