Bug#28898811 INCORRECTLY REJECT JOIN TO BE PUSHED DOWN
When a table was early eliminated as 'not maybe_pushable', setup of the
tables 'm_remap[]' matrix was also skipped. That had the side effect that
the table could not be referred in any later join conditions from *other* tables.

This patch change that logic such that the 'm_remap[]' matrix is set up
for all tables in a query, independent of whether the table itself could
be join-pushed down or not.