BUG #28817176 DOWNGRADE TESTS FAIL TO RESTART API NODE
Adds "/" at the end of base_dir(MYSQL_HOME environment variable) to be nicer to
MySQL tools.

ATRT_PID environment variable was corrupted when my_load_defaults() calls
my_search_option_files() for processing default options from argc & argv
where the previous environment variable MYSQL_HOME is checked whether it
ends with a '/' or not. If it doesn't a '/' was padded and the next
element in the array of environment variables was set to '\0' which in
turn deleted the entry of ATRT_PID. This is a bug in
${MYSQL_REPO}/mysys_ssl/my_default.cc and will be fixed with
coordination of server team. As a workaround, a '/' is padded at the end
of MYSQL_HOME in atrt itslef if not already present which skips the
check on the server side thereby generating a proper for ATRT_PID in the
list of environment variables and letting the NDB API node restart
itself appropriately.

This fix is part of the series of fixed done for downgrade tests that
have been failing due to a several reasons. One of which was improper
rolling restart of cluster which was fixed in the testcase
testDowngrade -n Upgrade_NR1 T1 as part of the WL#12233.
After this fix, downgrade tests from trunk to lower versions were still
failing during the restart of NDB API node. The reason for failue was
that the testcase was not able to find it's own process ID which was
stored as an environment variable ATRT_PID and ATRT_PID was missing from
the list of environment variables.