Bug#28922733 SPORADIC FAILURE OF NDB_DD_BASIC IN MTR WITH TABLE FULL ERROR (1114)
The test ndb_dd_basic fails sporadically with a table full error. This
table full error is actually triggered due to the tablespace getting
full. A tablespace ts1 is created at the beginning of the testfile and
the same is being used across the testfile by all the testcases. Even
though the test cases drop the tables after testing the queries, the
space won't be immediately freed, if an LCP is ongoing. The space will
be available to create new tables only after the LCP gets completed.
Due to this, we sometimes get the table is full error and a 'Out of
extents, tablespace full' warning along with it.

This patch fixes that by using seperate tablespaces for the individual
testcases. The tablespace along with the tables are cleaned up at the
end of each testcase. Also cleanup up the testcase wherever possible.

Change-Id: Iada10090d7abbe984cc276e9cd87bd0b786cfad6