Bug#27429615 CLUSTER: QUERY EXPLAIN SHOWS PUSHED CONDITIONS WHICH ARE NOT PUSHED
Bug fix:

Explain of a query executed with the table access type EQ_REF could also
show a condition being pushed down, even if condition pushdown is not
done for this access type.

It is of limited value to push a condition when only a single row
could be returned, so by design we do not want to do condition
pushdown in these cases.

The exception is if the table being accessed is part of a pushed join:
Then an entire branch in the joined query plan may be eliminated by
evaluating such a pushed condition. So we *do* support pushed condition
for pushed joins, independent of the access type being used.