Bug #28953158: DROP ROLE USERNAME SHOULD BE REJECTED
CREATE ROLE creates a user account with login disabled.
And it requires the CREATE ROLE privilege.
This differentiates it from CREATE USER that does create
a user account with login enabled. And it also requires
the CREATE USER privilege.
Same privilege (CREATE USER) is required for ALTER USER.
And for DROP USER too.
DROP ROLE requires the DROP USER or the DROP ROLE privileges.
For DROP ROLE privilege to be different from CREATE USER
the server shouldn't allow DROP ROLE to drop user accounts
if the session doesn't have CREATE USER and only has the
DROP ROLE privilege.
This will make the CREATE ROLE and DROP ROLE privileges a
pair that is safe to grant to somebody that you only want to
be creating roles and not user accounts.

So the fix here is to add an additional check to DROP ROLE
if the CREATE USER user privilege is not granted and the
DROP ROLE privilege is granted DROP ROLE should check if
any of the user accounts being dropped is a role (has login
enabled) and if so refuse to execute the statement.

Added a test case to cover this.