Bug#28867993: POSSIBLE ISSUE WITH MYSQL SERVER RESTART
Analysis
========
Creation of database with the same name as of redo log
files is allowed.

We currently do not restrict creation of databases with
the same name as redo log files. This may cause issues
during server start-up.

Fix
===
Creation of databases with names which are same as the redo
log file names are not allowed.

(cherry picked from commit 2a58dc71a18ba644ddbd63413fcc7ebeca74cdb0)