Bug#28813708 NDB_RESTORE LOOKS FOR NODE ID 0 IN BACKUP FILES  - *.CTL, *.DATA AND *.LOG
The ndb_restore tool uses the parameters --nodeid and --backupid to
construct filenames of files within the backup. If these parameters
are not provided, ndb_restore uses default values of 0, which are
invalid. So if either of these parameters is missing, an invalid
filename is constructed and ndb_restore fails with a file-open
error.

It is not safe to provide valid defaults for these parameters
since this may result in restoration of the wrong data or duplicate
data. So these parameters have been made mandatory for ndb_restore.