Bug#28923060 TEST_NDBINFO FAILS DUE TO UNEXPECTED NUMBER OF ROWS IN NDBINFO.CPUSTAT_1SEC
The ndbinfo.cpustat_1sec table has 20 measurement sets per thread, each
referencing a period of the named duration, to provide 20 seconds of
history for that particular duration. In the initial 20 seconds after
cluster has started, this history won't be there yet. Due to that the
table previously had garbage values sometimes. This has been fixed now
as a part of fix for Bug#28787472. testNdbinfo testsuite as a part of
its testcases checks the number of rows returned by various ndbinfo
tables in a loop to confirm thery are consistent. But due to the new
update mentioned above, cpustat_1sec table's total number of rows will
be increasing in the initial 20 seconds. The test reports this as an
error and fails. Other cpustat tables - cpustat_50ms and cpustat_20sec
also have similar issue.

Fixed the testcase by adjusting it to the new behaviour of cpustat
tables. Also removed an unused method from the test file.

Change-Id: Id6842ede7a9575ad3afdd5c99c93617d8df882e8